Dev merge into Main

+++ john_callahan_p0/src/main/java/com/revature/projectZero/service/ValidationService.java
@@ -3,17 +3,18 @@ package com.revature.projectZero.service;
 import com.revature.projectZero.pojos.Course;
 import com.revature.projectZero.pojos.Enrolled;
 import com.revature.projectZero.util.exceptions.InvalidRequestException;
-import com.revature.projectZero.util.exceptions.ResourcePersistenceException;
 import com.revature.projectZero.pojos.Faculty;
 import com.revature.projectZero.pojos.Student;
 import com.revature.projectZero.repositories.SchoolRepository;
-
+import com.revature.projectZero.util.exceptions.ResourcePersistenceException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import java.util.List;
 
 /**
-    A service-layer middleman that validates input before passing it to the database class responsible for
- placing it into the database.
-     */
+ *  A service-layer middleman that validates input before passing it to the database class responsible for
+ *  placing it into the database. It also does Session Tracking.
+ */
 
 public class ValidationService {
 
@@ -21,44 +22,48 @@ public class ValidationService {
 
         public ValidationService(SchoolRepository studentRepo){ this.schoolRepo = studentRepo; }
 
+        Logger logger = LogManager.getLogger(ValidationService.class);
         private boolean isValid = true;
         private Student authStudent;
         private Faculty authFac;
 
 
-        public void register(Student newStudent) throws Exception {
+        public void register(Student newStudent) {
+
                 if (!isUserValid(newStudent)) {
                         throw new InvalidRequestException("Invalid user data provided!");
-                }
 
-                if (schoolRepo.findStudentByUsername(newStudent.getUsername()) != null) {
-                        throw new ResourcePersistenceException("Provided username is already taken!");
                 }
-
                 schoolRepo.save(newStudent);
+                logger.info("New Student registered!" + newStudent);
+
         }
 
         // This will attempt to persist a created course to the courses database.
         public void createCourse(Course newCourse) {
-
                 try {
+                        // Verify that the class ID is not already taken.
+                        if (isClassIDTaken(newCourse.getClassID()))
+                                throw new InvalidRequestException("Class ID cannot already be taken!");
+
                         newCourse.setTeacher(this.authFac.getLastName());
-                        // TODO: Run this collected object against some criteria!
-                        schoolRepo.newCourse(newCourse);
-                } catch(Exception e) {
-                        System.out.println(e.getMessage());
+
+                        if (isCourseValid(newCourse)) {
+                                schoolRepo.newCourse(newCourse);
+                                logger.info("New Course created!" + newCourse);
+                        }
+                } catch(InvalidRequestException ire) {
+                        System.out.println(ire.getMessage());
                 }
         }
 
         // This enrolls a student into a course, grafting their username to it,
         // and placing it within the separate 'enrolled' database.
-        public void enroll(Course selectedCourse, String id) {
-
-                Enrolled enrollIn = new Enrolled(this.authStudent.getUsername(), selectedCourse.getName(), selectedCourse.getClassID(), selectedCourse.getDesc(), selectedCourse.getTeacher());
-                try {
+        public void enroll(Course selectedCourse) {
+                if (isCourseValid(selectedCourse)) {
+                        Enrolled enrollIn = new Enrolled(this.authStudent.getUsername(), selectedCourse.getName(),
+                                selectedCourse.getClassID(), selectedCourse.getDesc(), selectedCourse.getTeacher());
                         schoolRepo.enroll(enrollIn);
-                } catch(Exception e) {
-                        System.out.println(e.getMessage());
                 }
         }
 
@@ -69,7 +74,15 @@ public class ValidationService {
                 this.authFac = null;
         }
 
-        public Student login(String username, int hashPass) throws Exception {
+        public void setAuthStudent(Student authStudent) {
+                this.authStudent = authStudent;
+        }
+
+        public void setAuthFac(Faculty authFac) {
+                this.authFac = authFac;
+        }
+
+        public Student login(String username, int hashPass) {
                 if (username == null || username.trim().equals("")) {
                         throw new InvalidRequestException("Invalid user credentials provided!");
                 }
@@ -90,7 +103,7 @@ public class ValidationService {
                 }
         }
 
-        public Faculty facLogin(String username, int hashPass) throws Exception {
+        public Faculty facLogin(String username, int hashPass) {
                 if (username == null || username.trim().equals("")) {
                         throw new InvalidRequestException("Invalid user credentials provided!");
                 }
@@ -121,33 +134,67 @@ public class ValidationService {
         public List<Course> getTeacherClasses() { return schoolRepo.findCourseByTeacher(this.authFac.getLastName()); }
 
         public void updateCourse(Course newCourse, String id) {
-                try {
+                if (isCourseValid(newCourse)) {
                         String teacher = this.authFac.getLastName();
                         schoolRepo.updateCourse(newCourse, id, teacher);
-                } catch (Exception e) {
-                        System.out.println("Updated course not persisted! " + e.getMessage());
+                } else {
+                        throw new ResourcePersistenceException("Sorry, but that could not be persisted!");
                 }
         }
 
-        public void deleteCourse(String id) {
+        public boolean deleteCourse(String id) {
                 try {
-                        // TODO: Validate this input!
-                        schoolRepo.deleteCourse(id);
-                } catch (Exception e) {
-                        System.out.println(e.getMessage());
+                        if (isClassIDValid(id) && isClassIDTaken(id)) {
+                                schoolRepo.deleteCourse(id);
+                                logger.info("Course " + id + " deleted!");
+                                return true;
+                        }
+                } catch(InvalidRequestException ire) {
+                        System.out.println(ire.getMessage());
                 }
+                return false;
         }
 
-        public void deregister(String id) {
+        public boolean deregister(String id) {
                 try {
-                        // TODO: Validate this input!
-                        schoolRepo.deleteEnrolled(id, this.authStudent.getUsername());
-                } catch (Exception e) {
-                        System.out.println(e.getMessage());
+                        if (isClassIDValid(id) && isClassIDTaken(id)) {
+                                schoolRepo.deleteEnrolled(id, this.authStudent.getUsername());
+                                logger.info("Course " + id + " owned by " + authStudent.getUsername() + " deleted!");
+                                return true;
+                        }
+                } catch(InvalidRequestException ire) {
+                        System.out.println(ire.getMessage());
                 }
+                return false;
         }
 
-        // TODO: Implement some verification for Teacher-submitted courses!
+
+        public boolean isCourseValid(Course course) {
+                try {
+                        if (course == null) return false;
+                        if (course.getName() == null || course.getName().trim().equals(""))
+                                throw new InvalidRequestException("Course name cannot be null");
+                        if (course.getClassID() == null || course.getClassID().trim().equals(""))
+                                throw new InvalidRequestException("Class ID cannot be blank or null");
+                        if (course.getDesc() == null || course.getDesc().trim().equals(""))
+                                throw new InvalidRequestException("Course description cannot be blank or null");
+                        if (course.getTeacher() == null) throw new InvalidRequestException("Teacher cannot be null");
+
+                        isClassIDValid(course.getClassID());
+
+                        // Verify that the course description is up to par.
+                        if (course.getDesc().length() < 20 && course.getDesc().length() > 50)
+                                throw new InvalidRequestException("Course description must be at least twenty characters and no more than fifty characters");
+
+                        // Verify that the course name is within valid boundaries.
+                        if (course.getName().length() < 5 && course.getName().length() <= 20)
+                                throw new InvalidRequestException("Course name must be at least five characters and less than or equal to twenty characters");
+                } catch (InvalidRequestException ire) {
+                        System.out.println(ire.getMessage());
+                        return false;
+                }
+                return true;
+        }
 
         // This verifies that students are valid and fit to be placed in the system.
         public boolean isUserValid(Student user) {
@@ -157,7 +204,27 @@ public class ValidationService {
                 if (user.getEmail() == null || user.getEmail().trim().equals("")) return false;
                 if (user.getUsername() == null || user.getUsername().trim().equals("")) return false;
                 this.isUserUnique(user.getUsername());
+                this.isEmailUnique(user.getEmail());
                 return true;
+        }
+
+        public boolean isClassIDValid(String ClassID) {
+
+                // Verify that the class ID is exactly six characters long.
+                if (ClassID.length() != 6) throw new RuntimeException("Class ID must be exactly six characters long!");
+
+                // Verify that all characters in ClassID are uppercase
+                for (int i=0; i<ClassID.length(); i++) {
+                        if (Character.isLowerCase(ClassID.charAt(i))) {
+                                throw new InvalidRequestException("Class ID must be all capital letters!");
+                        }
+                }
+
+                return true;
+        }
+
+        private boolean isClassIDTaken(String ClassID) {
+                return schoolRepo.findCourseByID(ClassID) != null;
         }
 
         public boolean isUserUnique(String username) {
